Use Temporary Directory for MaveDB Score and Metadata Files
Creates a decorator for mapping routines which creates a temporary directory in which score set
metadata and score files can be downloaded. The directory path is then passed to the mapping routine so that these temp files
can be used by the mapper. Once the wrapped function exits, the temporary directory is purged.