Mock cdot data provider for tests
Although cdot fetching is not currently directly used in any tests,
the cdot REST data provider is imported into dcd_mapping.vrs_map.
The cdot data provider uses a ChainedSeqFetcher, to which specific
fasta file paths are passed in dcd_mapping.lookup.
This results in a FileNotFoundError during tests, since these fasta
files are not available outside of the mapper Docker container.
This fix uses a fake fasta file to generate a ChainedSeqFetcher,
so this change does not support actual cdot transcript fetches for
future tests.