Translate NT target sequence for targets with protein-level variants
If a target has only protein-level variants, but the provided target sequence
is a nucleotide sequence, translate the nucleotide sequence to an amino acid
sequence immediately after metadata ingestion.
This change avoids alignment errors that can occur when a target sequence has been
codon-optimized to a non-human organism. Since we do not have sufficient metadata
to assume that a target sequence has been codon-optimized, always perform translation
when there are no nucleotide-level variants for a target.