feat: compute gene info for all mapped targets.
Computes a new `gene_info` property for all mapped targets. This property is defined by an `hgnc_symbol` and a `selection_method`.

The hgnc symbol is the HGNC symbol of the gene to which this target relates. The selection method is the method by which this symbol was selected and may be:
- `tx_selection`: via the selected transcript
- `alignment_max_covered_bases`: based on the gene 'feature' (via Ensembl) which covered the most bases of the aligned target
- `variants_max_covered_bases`: same as `alignment_max_covered_bases`, but based on variant bases rather than aligned bases
- `target_metadata`: based on parsing the target metadata the user supplied
- `target_category`: no gene info was selected because the target was not protein coding (see #66)

Various helpers were added to `dcd_mapping.annotate` which support this calculation. Gene info selection should not cause job failures, and will simply fail to select gene info on failure.